backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=1.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
